+++ src/connectProvider/web3Provider/MEWconnect/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import MEWconnect from '../../../index';
 import { Transaction } from 'ethereumjs-tx';
 import WalletInterface from '../WalletInterface';
@@ -69,11 +70,14 @@ class MEWconnectWallet {
   }
 
   static setConnectionState(connectionState) {
-    if (!connectionState) MEWconnect.Initiator.connectionState = 'disconnected';
-    else MEWconnect.Initiator.connectionState = connectionState;
+    if (!connectionState) MEWconnect.Initiator.setConnectionState('disconnected');
+    else MEWconnect.Initiator.setConnectionState(connectionState);
+    debug('setConnectionState', MEWconnect.Initiator.connectionState)
   }
 
   static getConnectionState() {
+    debug('getConnectionState', MEWconnect.Initiator.connectionState)
+
     if (!MEWconnect.Initiator.connectionState) return 'disconnected';
     return MEWconnect.Initiator.connectionState;
   }