+++ src/connectWindow/popupWindowDesign.js
@@ -211,26 +211,20 @@ const cssStyles = `
         }
     `;
 
-const htmlDesign = (refresh, image, playStore, appStore, camera) => {
+const htmlDesign = (refresh, image, playStore, appStore, camera, iconImage) => {
   return `
-      <html>
-      <head>
-        <meta charset="utf-8"/>
-        <meta name="viewport" content="width=device-width,initial-scale=1,user-scalable=no"/>
-        <meta name="theme-color" content="#000000"/>
-        <title>MEWconnect</title>
-      </head>
-
-<body>
     <div class="outer-container">
       <div class="container">
       <div class="upper-text">
+        <div class="close-mew-modal" id="close-mew-modal" aria-label="close modal" data-close>
+          <img src="${iconImage}" height="15" width="11"/>
+        </div>
         <p class="text-one">Connect to MEW wallet app</p>
         <p class="text-two">Scan this code to connect</p>
        </div>
 
         <div class="qr-code">
-          <canvas id="canvas"></canvas>
+          <canvas id="canvas-for-mewconnect-qr-code"></canvas>
         </div>
         <div id="refresh-container" class="refreshIcon hidden">
           <img id="refresh" src="${refresh}" />
@@ -283,22 +277,6 @@ const htmlDesign = (refresh, image, playStore, appStore, camera) => {
         </div>
       </div>
     </div>
-  <script>
-  const channel = new BroadcastChannel('refresh-channel');
-  const refreshContainer = window.document.getElementById("refresh-container")
-  const refreshButton = window.document.getElementById("refresh");
-
-  refreshButton.addEventListener("click", () => {
-    channel.postMessage("refresh");
-  })
-
-  // setTimeout(() => {
-  //   refreshContainer.className = refreshContainer.className.replace('hidden', '');
-  // }, 5000)
-
-  </script>
-</body>
-</html>
 
 `;
 };
@@ -654,4 +632,125 @@ const windowInformer = spaceman => {
       </div>
 `;
 };
-export { cssStyles, htmlDesign, noticetext, windowInformer };
+
+const modalFrame = innerContent => {
+  return `
+    <div class="mew-wallet-modal is-visible" id="mew-wallet-modal"></div>
+    <div class="mew-wallet-modal-container is-visible" id="mew-wallet-modal-container">
+      <div class="modal-dialog is-visible" id="mew-mobile-modal-dialog">
+        <section class="modal-content">
+        ${innerContent}
+        </section>
+      </div>
+    </div>
+`;
+};
+
+const modalCSS = (additionalCss = '') => {
+  return `
+${additionalCss}
+
+      div.close-mew-modal{
+        position: relative;
+        padding-top: 10px;
+        left: 100%;
+        width: 20px;
+        cursor: pointer;
+      }
+      
+      .mew-wallet-modal {
+        position: fixed;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        padding: 1rem;
+        background: black;
+        cursor: default;
+        visibility: hidden;
+        opacity: 0;
+        transition: all 0.35s ease-in;
+      }
+
+      .mew-wallet-modal.is-visible {
+        visibility: visible;
+        opacity: 0.25;
+        z-index: 999999;
+      }
+
+      div.modal-dialog {
+        background: rgb(255, 255, 255);
+        border-radius: 16px;
+        box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.05),
+        0px 3px 6px 0px rgba(0, 0, 0, 0.05),
+        0px 8px 16px 0px rgba(0, 0, 0, 0.05);
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        position: relative;
+        min-width: 448px;
+        max-width: 448px;
+        min-height: 558px;
+        overflow: auto;
+        opacity: 0;
+        visibility: hidden;
+        z-index: 999999;
+      }
+
+      .mew-wallet-modal-container {
+        position: fixed;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        padding: 1rem;
+        background-color: transparent;
+        cursor: default;
+        visibility: hidden;
+        opacity: 0;
+        transition: all 0.35s ease-in;
+        z-index: 999999;
+      }
+
+      div.mew-wallet-modal-container.is-visible {
+        visibility: visible;
+        opacity: 1;
+        background-color: transparent;
+      }
+
+      div.modal-dialog.is-visible {
+        visibility: visible;
+        opacity: 1;
+        z-index: 99999999999999;
+      }
+
+      .modal-dialog > * {
+        padding: 1rem;
+      }
+
+      .modal-header {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+      }
+
+      .modal-header .close-modal {
+        font-size: 1.5rem;
+      }
+
+`;
+};
+export {
+  cssStyles,
+  htmlDesign,
+  noticetext,
+  windowInformer,
+  modalFrame,
+  modalCSS
+};